+++ src/søknader/felles/steg/1-omdeg/OmDegContext.tsx
@@ -72,6 +72,11 @@ const [OmDegProvider, useOmDeg] = constate(
 
     const [medlemskap, settMedlemskap] = useState(søknad.medlemskap);
     const [sivilstatus, settSivilstatus] = useState(søknad.sivilstatus);
+    const [søkerBorPåRegistrertAdresse, settSøkerBorPåRegistrertAdresse] =
+      useState(søknad.søkerBorPåRegistrertAdresse);
+    const [harMeldtAdresseendring, settHarMeldtAdresseendring] = useState(
+      søknad.adresseopplysninger?.harMeldtAdresseendring
+    );
 
     useEffect(() => {
       if (mellomlagretSøknad?.søknad.medlemskap) {
@@ -79,6 +84,18 @@ const [OmDegProvider, useOmDeg] = constate(
       }
       if (mellomlagretSøknad?.søknad.sivilstatus) {
         settSivilstatus(mellomlagretSøknad.søknad.sivilstatus);
+      }
+      if (mellomlagretSøknad?.søknad.søkerBorPåRegistrertAdresse) {
+        settSøkerBorPåRegistrertAdresse(
+          mellomlagretSøknad.søknad.søkerBorPåRegistrertAdresse
+        );
+      }
+      if (
+        mellomlagretSøknad?.søknad.adresseopplysninger?.harMeldtAdresseendring
+      ) {
+        settHarMeldtAdresseendring(
+          mellomlagretSøknad.søknad.adresseopplysninger.harMeldtAdresseendring
+        );
       }
     }, [mellomlagretSøknad]);
 
@@ -95,6 +112,10 @@ const [OmDegProvider, useOmDeg] = constate(
     };
 
     return {
+      søkerBorPåRegistrertAdresse,
+      settSøkerBorPåRegistrertAdresse,
+      harMeldtAdresseendring,
+      settHarMeldtAdresseendring,
       sivilstatus,
       settSivilstatus,
       medlemskap,
