+++ src/søknader/felles/steg/1-omdeg/personopplysninger/Personopplysninger.tsx
@@ -9,7 +9,6 @@ import {
 } from './PersonopplysningerConfig';
 import { hentBooleanFraValgtSvar } from '../../../../../utils/spørsmålogsvar';
 import { ISpørsmål, ISvar } from '../../../../../models/felles/spørsmålogsvar';
-import { Søker } from '../../../../../models/søknad/person';
 import { ISpørsmålBooleanFelt } from '../../../../../models/søknad/søknadsfelter';
 import { Stønadstype } from '../../../../../models/søknad/stønadstyper';
 import { useLokalIntlContext } from '../../../../../context/LokalIntlContext';
@@ -17,18 +16,9 @@ import AlertStripeDokumentasjon from '../../../../../components/AlertstripeDokum
 import { hentTekst } from '../../../../../utils/søknad';
 import { OmDeg } from './OmDeg';
 import { Alert, VStack } from '@navikt/ds-react';
+import { useOmDeg } from '../OmDegContext';
 
 interface Props {
-  søker: Søker;
-  settDokumentasjonsbehov: (
-    spørsmål: ISpørsmål,
-    valgtSvar: ISvar,
-    erHuketAv?: boolean
-  ) => void;
-  søkerBorPåRegistrertAdresse?: ISpørsmålBooleanFelt;
-  settSøkerBorPåRegistrertAdresse: (
-    søkerBorPåRegistrertAdresse: ISpørsmålBooleanFelt
-  ) => void;
   harMeldtAdresseendring?: ISpørsmålBooleanFelt;
   settHarMeldtAdresseendring: (
     harMeldtAdresseendring: ISpørsmålBooleanFelt
@@ -37,15 +27,18 @@ interface Props {
 }
 
 const Personopplysninger: React.FC<Props> = ({
-  søker,
-  settDokumentasjonsbehov,
-  søkerBorPåRegistrertAdresse,
-  settSøkerBorPåRegistrertAdresse,
   harMeldtAdresseendring,
   settHarMeldtAdresseendring,
   stønadstype,
 }) => {
   const intl = useLokalIntlContext();
+  const {
+    søknad,
+    settSøknad,
+    settDokumentasjonsbehov,
+    søkerBorPåRegistrertAdresse,
+  } = useOmDeg();
+  const { søker } = søknad.person;
 
   const settPersonopplysningerFelt = (
     spørsmål: ISpørsmål,
@@ -57,6 +50,18 @@ const Personopplysninger: React.FC<Props> = ({
       svarid: valgtSvar.id,
       label: hentTekst(spørsmål.tekstid, intl),
       verdi: svar,
+    });
+  };
+
+  const settSøkerBorPåRegistrertAdresse = (
+    søkerBorPåRegistrertAdresse: ISpørsmålBooleanFelt
+  ) => {
+    settSøknad((prevSoknad: any) => {
+      return {
+        ...prevSoknad,
+        adresseopplysninger: undefined,
+        søkerBorPåRegistrertAdresse: søkerBorPåRegistrertAdresse,
+      };
     });
   };
